wip v2 - shaded knobs
