+++ src/Intermix.cpp
@@ -1,6 +1,7 @@
 #include "plugin.hpp"
 #include "digital.hpp"
 #include "IntermixBase.hpp"
+#include "components/Knobs.hpp"
 #include "components/MatrixButton.hpp"
 
 namespace StoermelderPackOne {
