+++ src/Pile.cpp
@@ -1,4 +1,5 @@
 #include "plugin.hpp"
+#include "components/Knobs.hpp"
 #include "components/VoltageLedDisplay.hpp"
 
 namespace StoermelderPackOne {
