+++ src/PilePoly.cpp
@@ -1,4 +1,5 @@
 #include "plugin.hpp"
+#include "components/Knobs.hpp"
 
 namespace StoermelderPackOne {
 namespace PilePoly {
