+++ src/ReMove.cpp
@@ -1,5 +1,6 @@
 #include "plugin.hpp"
 #include "MapModuleBase.hpp"
+#include "components/Knobs.hpp"
 #include <thread>
 #include <random>
